Mark API private and fix some inconsistent stuff
