+++ hearth/lib/hearth.rb
@@ -11,7 +11,8 @@ require_relative 'hearth/middleware_builder'
 require_relative 'hearth/middleware_stack'
 require_relative 'hearth/number_helper'
 require_relative 'hearth/output'
-require_relative 'hearth/query'
+require_relative 'hearth/query/param'
+require_relative 'hearth/query/param_list'
 require_relative 'hearth/retry'
 require_relative 'hearth/structure'
 require_relative 'hearth/stubbing/client_stubs'
@@ -19,7 +20,6 @@ require_relative 'hearth/stubbing/stubs'
 require_relative 'hearth/time_helper'
 require_relative 'hearth/union'
 require_relative 'hearth/validator'
-require_relative 'hearth/waiters/errors'
 require_relative 'hearth/waiters/poller'
 require_relative 'hearth/waiters/waiter'
 require_relative 'hearth/xml'
