--- hearth/lib/hearth/query.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'query/param'
-require_relative 'query/param_list'
